oidc_server/requirements.txt
Flask
Flask-SQLAlchemy
Authlib==0.13
python-dotenv
mysqlclient==2.1.1